+++ crates/types/src/signaling/control/event_info.rs
@@ -2,6 +2,8 @@
 //
 // SPDX-License-Identifier: EUPL-1.2
 
+use crate::core::EventId;
+
 #[allow(unused_imports)]
 use crate::imports::*;
 
@@ -9,6 +11,8 @@ use crate::imports::*;
 #[derive(Clone, Debug, PartialEq, Eq)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub struct EventInfo {
+    /// The id of the event
+    pub id: EventId,
     /// The title of the event
     pub title: String,
 }